feat: スイッチルールの型を追加
